Removing the core subpackage
