+++ bundle/src/main/java/com/adobe/acs/commons/adobeio/service/package-info.java
@@ -17,25 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.adobeio.core.service.impl;
-
 /**
- * List of constants used in the Adobe I/O connection
+ * Adobe I/O services.
  */
-public final class AdobeioConstants {
-
-   public static final String CONTENT_TYPE_APPLICATION_JSON = "application/json";
-
-   public static final String X_API_KEY = "X-Api-Key";
-
-   public static final String JSON_ACCESS_TOKEN = "access_token";
-   public static final String JK_PKEY = "PKey";
-   public static final String JK_SUBSCRIBER = "subscriber";
-
-   public static final String RESULT_ERROR = "error";
-   public static final String RESULT_NO_DATA = "nodata";
-
-   private AdobeioConstants() {
-   }
-
-}
+package com.adobe.acs.commons.adobeio.service;
